fix compile time warnings

+++ sdk/test/metrics/sketch_aggregator_test.cc
@@ -58,7 +58,7 @@ TEST(Sketch, NormalValues)
   EXPECT_EQ(alpha.get_counts(), correct);
 
   std::vector<double> captured_bounds = alpha.get_boundaries();
-  for (int i = 0; i < captured_bounds.size(); i++)
+  for (size_t i = 0; i < captured_bounds.size(); i++)
   {
     captured_bounds[i] = round(captured_bounds[i]);
   }
